Move `gen_helpers` into a module of `eth2spec` package
