+++ tests/generators/bls/requirements.txt
@@ -1,4 +1,2 @@
-py_ecc==5.1.0
-eth-utils==1.6.0
-../../core/gen_helpers
+pytest>=4.4
 ../../../
