+++ tests/generators/ssz_generic/requirements.txt
@@ -1,3 +1,2 @@
-eth-utils==1.6.0
-../../core/gen_helpers
+pytest>=4.4
 ../../../
